src/requirements/base.txt
Django==1.6.8
Fabric==1.9.0
git+git://github.com/tucarga/django-registration-rest-framework@master
django-filter==0.8
South==1.0.1
argparse==1.2.1
async==0.6.1
django-suit==0.2.11
django-extensions==1.4.4
django-ipware==0.1.0
djangorestframework==2.4.3
ecdsa==0.11
gitdb==0.5.4
httplib2==0.9
paramiko==1.15.1
psycopg2==2.5.4
pycrypto==2.6.1
pyyoutrack==1.0.6
raven==5.1.1
six==1.8.0
smmap==0.8.2
urllib2-file==0.2.1
wsgiref==0.1.2
django-allauth==0.18.0
django-cors-headers==0.12
Pillow==2.6.1
easy-thumbnails==2.2
uwsgi
Werkzeug
git+git://github.com/tucarga/django-registration-rest-framework@master
django_countries
drf-nested-routers
django_filters
django-import-export==0.2.6
django_deliver_notifications==0.1.2
django-hstore
djangorestframework-hstore
celery
django-celery
django-reversion
weasyprint
elasticsearch
django-haystack
pyelasticsearch
django-app-metrics